docs: modify README.md to test COMMIT app
This is a trival commit. I've simply added an emoji to the README.md to
see if it support emojis, it was added as unicode hex. But most
importantly this commit is to test the capabilities of the COMMIT app. I
want to see if it enforces line length among other things.